minor cosmetic fix
git-svn-id: http://svn.xiph.org/trunk/speex@4363 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/speexenc.c
@@ -649,7 +649,7 @@ int main(int argc, char **argv)
          cumul_bits += tmp;
          enc_frames += 1;
          if (vad_enabled || vbr_enabled || abr_enabled)
-            fprintf (stderr, "Bitrate is use: %d bps  (average %d bps)   ", tmp, cumul_bits/enc_frames);
+            fprintf (stderr, "Bitrate is use: %d bps  (average %d bps)   ", tmp, (int)(cumul_bits/enc_frames));
          else
             fprintf (stderr, "Bitrate is use: %d bps     ", tmp);
